Point Issue Reporting to xctest-dynamic-overlay repo
Swift Package Manager honors redirects, but it appears to associate the
path suffix with the package name, and this conflicts with package
resolution in certain (but not all) cases.

I think we have no choice but to roll back everything to point to the
original xctest-dynamic-overlay URL and extract Issue Reporting to a
dedicated repo.